添加登录协议/去除列表设备复制功能

+++ wallet/user.ts
@@ -64,7 +64,7 @@ async function tryLogin(state: LoginState, key2?: string, ref?: string) { // tes
 		await sdk.user.methods.setUser({ key2, ref }, { signer: new MySigner(state) }); // check access and set key2
 	} catch (err: any) { // ILLEGAL ACCESS
 		if (err.errno == errno.ERR_ILLEGAL_ACCESS[0]) {
-			if (location.pathname != '/login') {
+			if (!['/agreement', '/login'].includes(location.pathname)) {
 				console.log(state, "state", key2, "key2", '过期数据');
 				logout();
 				throw new Error('身份验证过期!');
@@ -91,6 +91,7 @@ export function loginState() {
 }
 
 export function logout() {
+	if (['/agreement', '/login'].includes(location.pathname)) return false;
 	useTouristState(); // tourist user
 	storage.delete('loginState');
 	var nav = Root.current.refs.nav as Nav;
